pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clibroker"
version = "0.1.0"
description = "Policy-driven CLI command broker — secure REST gateway for wrapped CLI tools"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "mcp>=1.6",
]

[project.optional-dependencies]
client = [
    "httpx>=0.27",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "asgi-lifespan>=2.1",
    "anyio>=4.0",
]

[project.scripts]
clibroker = "clibroker.__main__:main"
clibroker-client = "clibroker.client.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/clibroker"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
